hiresignal-proto/backend/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.30.0
celery[redis]==5.4.0
redis==5.0.4
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.3.0
email-validator==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
sentence-transformers==2.7.0
scikit-learn==1.5.0
spacy==3.7.4
pdfplumber==0.11.1
python-docx==1.1.2
pytesseract==0.3.13
pdf2image==1.17.0
langdetect==1.0.9
pyahocorasick==2.0.0
openpyxl==3.1.2
python-magic==0.4.27
chardet==5.2.0
httpx==0.27.0
websockets==12.0
python-dotenv==1.0.0
